+++ src/components/experience-card.tsx
@@ -1,0 +1,49 @@
+"use client";
+import Link from "next/link";
+
+export interface ExperienceCardProps {
+  role: string;
+  company: string;
+  companyUrl: string;
+  startDate: string;
+  endDate: string;
+  highlight: string;
+  description: string;
+}
+
+export function ExperienceCard({
+  role,
+  company,
+  companyUrl,
+  startDate,
+  endDate,
+  highlight,
+  description,
+}: ExperienceCardProps) {
+  return (
+    <div>
+      {/* Header */}
+      <div className="mb-2 text-foreground">
+        <p className="font-medium">
+          {role} @{" "}
+          <Link
+            href={companyUrl}
+            target="_blank"
+            rel="noopener noreferrer"
+            className="underline"
+          >
+            {company}
+          </Link>
+        </p>
+        <p className="text-xs text-muted-foreground/90">
+          {startDate} &mdash; {endDate}
+        </p>
+      </div>
+
+      {/* Description */}
+      <p className="text-sm sm:text-[15px]">
+        {highlight} {description}
+      </p>
+    </div>
+  );
+}
